+++ src/components/date-time-input/date-time-input.spec.ts
@@ -60,25 +60,27 @@ describe('Date Time Input component', () => {
       await elementUpdated(el);
       expect(el.placeholder).to.equal('dd.MM.yyyy');
       expect(el.inputFormat).to.equal('dd.MM.yyyy');
+      expect(el.displayFormat).to.equal('d.M.yyyy');
     });
 
     it('should update inputFormat with value according to locale', async () => {
       el.value = new Date(2020, 2, 3);
       await elementUpdated(el);
-      expect(input.value).to.equal('03/03/2020');
+      expect(input.value).to.equal('3/3/2020');
 
       el.locale = 'no';
       await elementUpdated(el);
       expect(el.placeholder).to.equal('dd.MM.yyyy');
       expect(el.inputFormat).to.equal('dd.MM.yyyy');
-      expect(input.value).to.equal('03.03.2020');
+      expect(el.displayFormat).to.equal('d.M.yyyy');
+      expect(input.value).to.equal('3.3.2020');
     });
 
     it('should use inputFormat if no displayFormat is defined - issue 1114', async () => {
       el.inputFormat = 'yyyy#MM#dd';
       await elementUpdated(el);
 
-      expect(el.displayFormat).to.be.undefined;
+      expect((el as any)._displayFormat).to.be.undefined;
       expect(input.placeholder).to.equal('yyyy#MM#dd');
 
       el.value = new Date(2020, 2, 3);
@@ -103,12 +105,12 @@ describe('Date Time Input component', () => {
     });
 
     it('should use displayFormat when defined', async () => {
-      expect(el.displayFormat).to.be.undefined;
+      expect((el as any)._displayFormat).to.be.undefined;
 
       el.value = new Date(2020, 2, 3);
       await elementUpdated(el);
 
-      expect(input.value).to.equal('03/03/2020');
+      expect(input.value).to.equal('3/3/2020');
 
       el.displayFormat = 'dd.MM/yyyy';
       await elementUpdated(el);
@@ -154,9 +156,17 @@ describe('Date Time Input component', () => {
       el.value = new Date(2020, 9, 12, 12);
       el.displayFormat = 'd.MM hh:mm ttt';
       await elementUpdated(el);
-      expect(input.value).to.equal('12.10 12:00 noon');
+      expect(input.value).to.equal('12.10 12:00 PM');
 
       el.displayFormat = 'd.MM hh:mm ttttt';
+      await elementUpdated(el);
+      expect(input.value).to.equal('12.10 12:00 p');
+
+      el.displayFormat = 'd.MM hh:mm bbbb';
+      await elementUpdated(el);
+      expect(input.value).to.equal('12.10 12:00 noon');
+
+      el.displayFormat = 'd.MM hh:mm bbbbb';
       await elementUpdated(el);
       expect(input.value).to.equal('12.10 12:00 n');
     });
@@ -192,7 +202,7 @@ describe('Date Time Input component', () => {
       await elementUpdated(el);
 
       expect(eventSpy).calledWith('igcChange');
-      expect(input.value).to.deep.equal('01/01/2000');
+      expect(input.value).to.deep.equal('1/1/2000');
     });
 
     it('should correctly switch between different pre-defined date formats', async () => {
@@ -200,7 +210,7 @@ describe('Date Time Input component', () => {
 
       let formattedDate = setDateFormat('short', targetDate);
 
-      expect(el.displayFormat).to.be.undefined;
+      expect((el as any)._displayFormat).to.be.undefined;
 
       el.value = new Date(2020, 2, 3);
       el.displayFormat = 'short';
@@ -266,7 +276,7 @@ describe('Date Time Input component', () => {
     it('should clear input date on clear', async () => {
       el.value = new Date(2020, 2, 3);
       await elementUpdated(el);
-      expect(input.value).to.equal('03/03/2020');
+      expect(input.value).to.equal('3/3/2020');
 
       el.clear();
       await elementUpdated(el);
@@ -657,7 +667,7 @@ describe('Date Time Input component', () => {
 
     it('non filled parts have default value set on blur', async () => {
       el.inputFormat = 'dd.MM.yyyy';
-      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat);
+      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat, 'en');
 
       el.focus();
       await elementUpdated(el);
@@ -708,7 +718,7 @@ describe('Date Time Input component', () => {
 
     it('set value when input is complete', async () => {
       el.inputFormat = 'dd.MM.yyyy';
-      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat);
+      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat, 'en');
 
       el.focus();
       await elementUpdated(el);
@@ -737,7 +747,7 @@ describe('Date Time Input component', () => {
 
     it('set value to null when input is complete and invalid', async () => {
       el.inputFormat = 'dd.MM.yyyy';
-      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat);
+      const parts = DateTimeUtil.parseDateTimeFormat(el.inputFormat, 'en');
 
       el.focus();
       await elementUpdated(el);
@@ -822,7 +832,7 @@ describe('Date Time Input component', () => {
     it('Drop behavior', async () => {
       el.value = new Date(2020, 2, 3);
       await elementUpdated(el);
-      expect(input.value).to.equal('03/03/2020');
+      expect(input.value).to.equal('3/3/2020');
 
       input.value = '1010';
       input.setSelectionRange(0, 4);
@@ -1092,10 +1102,7 @@ describe('Date Time Input component', () => {
 
     describe('Validation', () => {
       const spec = createFormAssociatedTestBed<IgcDateTimeInputComponent>(html`
-        <igc-date-time-input
-          name="date-time"
-          .defaultValue=${null}
-        ></igc-date-time-input>
+        <igc-date-time-input name="date-time"></igc-date-time-input>
       `);
 
       beforeEach(async () => {