+++ src/components/mask-input/validators.ts
@@ -1,4 +1,4 @@
-import { validationResourcesKeys } from '../common/i18n/utils.js';
+import { ValidationResourceStringsEN } from 'igniteui-i18n-core';
 import { requiredValidator, type Validator } from '../common/validators.js';
 import type IgcMaskInputComponent from './mask-input.js';
 
@@ -6,7 +6,7 @@ export const maskValidators: Validator<IgcMaskInputComponent>[] = [
   requiredValidator,
   {
     key: 'badInput',
-    messageResourceKey: validationResourcesKeys.mask,
+    message: ValidationResourceStringsEN.mask_validation_error!,
     // @ts-expect-error - protected access
     isValid: ({ _parser, _maskedValue }) => _parser.isValidString(_maskedValue),
   },