+++ src/components/ripple/ripple.ts
@@ -1,7 +1,10 @@
-import { html, LitElement } from 'lit';
-
+import { LitElement, nothing } from 'lit';
 import { registerComponent } from '../common/definitions/register.js';
-import { addSafeEventListener } from '../common/util.js';
+import {
+  addSafeEventListener,
+  getScaleFactor,
+  setStyles,
+} from '../common/util.js';
 import { styles } from './ripple.material.css.js';
 
 const rippleFrames: Keyframe[] = [
@@ -17,7 +20,7 @@ const rippleAnimation: KeyframeAnimationOptions = {
 
 let rippleElement: HTMLElement;
 
-function getRippleElement() {
+function getRippleElement(): HTMLSpanElement {
   if (!rippleElement) {
     rippleElement = document.createElement('span');
   }
@@ -41,12 +44,19 @@ export default class IgcRippleComponent extends LitElement {
 
   constructor() {
     super();
-    addSafeEventListener(this, 'pointerdown', this.handler);
+    addSafeEventListener(this, 'pointerdown', this._handler);
   }
 
-  private handler = ({ clientX, clientY }: PointerEvent) => {
+  private async _handler(event: PointerEvent): Promise<void> {
+    if (event.button !== 0) {
+      return;
+    }
+
     const element = getRippleElement();
-    const { radius, top, left } = this.getDimensions(clientX, clientY);
+    const { radius, top, left } = this._getDimensions(
+      event.clientX,
+      event.clientY
+    );
 
     const styles: Partial<CSSStyleDeclaration> = {
       position: 'absolute',
@@ -55,8 +65,8 @@ export default class IgcRippleComponent extends LitElement {
       transformOrigin: 'center',
       transform: 'translate3d(0, 0, 0) scale(0)',
       willChange: 'opacity, transform',
-      margin: '0 !important',
-      border: 'none !important',
+      margin: '0',
+      border: 'none',
       width: `${radius}px`,
       height: `${radius}px`,
       borderRadius: '50%',
@@ -65,28 +75,28 @@ export default class IgcRippleComponent extends LitElement {
       background: 'var(--color, var(--ig-gray-800))',
     };
 
-    Object.assign(element.style, styles);
+    setStyles(element, styles);
     this.renderRoot.appendChild(element);
 
-    element
-      .animate(rippleFrames, rippleAnimation)
-      .finished.then(() => element.remove());
-  };
+    await element.animate(rippleFrames, rippleAnimation).finished;
+    element.remove();
+  }
 
-  private getDimensions(x: number, y: number) {
+  private _getDimensions(x: number, y: number) {
     const rect = this.getBoundingClientRect();
+    const factor = getScaleFactor(this);
     const radius = Math.max(rect.width, rect.height);
     const halfRadius = radius / 2;
 
     return {
       radius,
-      top: Math.round(y - rect.top - halfRadius),
-      left: Math.round(x - rect.left - halfRadius),
+      top: Math.round((y - rect.top) * factor.y - halfRadius),
+      left: Math.round((x - rect.left) * factor.x - halfRadius),
     };
   }
 
   protected override render() {
-    return html``;
+    return nothing;
   }
 }